+++ R/utils.R
@@ -94,11 +94,6 @@ get_folder_datetime = function(folder, verbose=TRUE){
 }
 
 
-can_be_numeric = function(x){
-  stopifnot(is.atomic(x) || is.list(x))
-  xnum_na <- suppressWarnings(is.na(as.numeric(x)))
-  all(is.na(x)==xnum_na)
-}
 
 #' @noRd
 #' @keywords internal
@@ -113,18 +108,7 @@ format_ymdhm = function(x){
   format(x, "%Y-%m-%d %Hh%M")
 }
 
-#' @noRd
-#' @keywords internal
-is.Date = function (x) {
-  inherits(x, "POSIXt") || inherits(x, "POSIXct") || inherits(x, "Date")
-}
 
-#' @noRd
-#' @keywords internal
-#' @source https://stackoverflow.com/a/57261396/3888000
-is_invalid_utf8 = function(x){
-  !is.na(x) & is.na(iconv(x, "UTF-8", "UTF-8"))
-}
 
 #' @noRd
 #' @keywords internal
@@ -133,34 +117,6 @@ is_invalid_utf8 = function(x){
 `%0%` = function (x, y) {
   if(length(x) == 0L) y else x
 }
-
-#' @noRd
-#' @keywords internal
-#' @importFrom cli cli_warn
-#' @importFrom dplyr arrange desc filter mutate
-#' @importFrom glue glue
-#' @importFrom rlang set_names
-#' @importFrom tidyr unnest
-check_invalid_utf8 = function(lookup=get_lookup(), warn=FALSE){
-  stopifnot(!is.null(lookup))
-  x = lookup %>% 
-    arrange(desc(nrow)) %>% 
-    unnest(c(names, labels)) %>% 
-    filter(is_invalid_utf8(labels)) %>% 
-    mutate(
-      dataset, names, labels, 
-      valid_labels=iconv(labels, to="UTF-8"),
-      .keep = "none"
-    )
-  
-  bad_utf8 = glue("{x$dataset}${x$names} ({x$valid_labels}) ") %>% set_names("i")
-  if(nrow(x)>0 && isTRUE(warn)){
-    cli_warn(c("Found {length(bad_utf8)} invalid UTF-8 label{?s}:", bad_utf8))
-  }
-  
-  x
-}
-
 
 #' any_of() with case sensitivity
 #' @noRd